Fix asan failure
