+++ dbms/src/Storages/KVStore/FFI/tests/gtest_status_server.cpp
@@ -13,9 +13,16 @@
 // limitations under the License.
 
 #include <Common/FailPoint.h>
+#include <Common/StringUtils/StringRefUtils.h>
 #include <Databases/DatabaseTiFlash.h>
 #include <Debug/MockKVStore/MockUtils.h>
 #include <Interpreters/Context.h>
+#include <Interpreters/InterpreterCreateQuery.h>
+#include <Interpreters/InterpreterDropQuery.h>
+#include <Parsers/ASTCreateQuery.h>
+#include <Parsers/ASTDropQuery.h>
+#include <Parsers/ParserCreateQuery.h>
+#include <Parsers/parseQuery.h>
 #include <Storages/IManageableStorage.h>
 #include <Storages/KVStore/FFI/ProxyFFI.h>
 #include <Storages/KVStore/FFI/ProxyFFICommon.h>
@@ -77,6 +84,13 @@ public:
         p = path + "/data/";
         TiFlashTestEnv::tryRemovePath(p, /*recreate=*/true);
     }
+
+    static void releaseResp(const EngineStoreServerHelper & helper, HttpRequestRes && res)
+    {
+        // release all raw cpp ptr in HttpRequestRes
+        helper.fn_gc_raw_cpp_ptr(res.res.inner.ptr, res.res.inner.type);
+        helper.fn_gc_raw_cpp_ptr(res.api_name.inner.ptr, res.api_name.inner.type);
+    }
 };
 
 TEST_F(StatusServerTest, TestSyncSchema)
@@ -108,7 +122,7 @@ try
         EXPECT_EQ(res.status, HttpRequestStatus::Ok);
         // normal errmsg is nil.
         EXPECT_EQ(res.res.view.len, 0);
-        delete (static_cast<RawCppString *>(res.res.inner.ptr));
+        releaseResp(helper, std::move(res));
     }
 
     {
@@ -122,7 +136,7 @@ try
         EXPECT_EQ(res.status, HttpRequestStatus::Ok) << magic_enum::enum_name(res.status);
         // normal errmsg is nil.
         EXPECT_EQ(res.res.view.len, 0);
-        delete (static_cast<RawCppString *>(res.res.inner.ptr));
+        releaseResp(helper, std::move(res));
     }
 
     {
@@ -137,7 +151,7 @@ try
         EXPECT_EQ(res_err.status, HttpRequestStatus::InternalError) << magic_enum::enum_name(res_err.status);
         StringRef sr(res_err.res.view.data, res_err.res.view.len);
         EXPECT_EQ(sr.toString(), "{\"errMsg\":\"sync schema failed\"}");
-        delete (static_cast<RawCppString *>(res_err.res.inner.ptr));
+        releaseResp(helper, std::move(res_err));
     }
 
     // test sync schema failed
@@ -152,7 +166,7 @@ try
         EXPECT_EQ(res_err1.status, HttpRequestStatus::InternalError) << magic_enum::enum_name(res_err1.status);
         StringRef sr(res_err1.res.view.data, res_err1.res.view.len);
         EXPECT_EQ(sr.toString(), "{\"errMsg\":\"Fail point FailPoints::sync_schema_request_failure is triggered.\"}");
-        delete (static_cast<RawCppString *>(res_err1.res.inner.ptr));
+        releaseResp(helper, std::move(res_err1));
     }
 
     dropDataBase("db_1");
@@ -184,7 +198,7 @@ try
         EXPECT_NE(res.res.view.len, 0);
         // normal response contains store_status ok info
         EXPECT_NE(std::string_view(res.res.view.data).find("[+]store_status ok"), std::string_view::npos);
-        delete (static_cast<RawCppString *>(res.res.inner.ptr));
+        releaseResp(helper, std::move(res));
     }
 
     {
@@ -202,8 +216,199 @@ try
         EXPECT_NE(res.res.view.len, 0);
         // error response contains store_status fail info
         EXPECT_NE(std::string_view(res.res.view.data).find("[-]store_status fail:"), std::string_view::npos);
-        delete (static_cast<RawCppString *>(res.res.inner.ptr));
+        releaseResp(helper, std::move(res));
     }
+}
+CATCH
+
+
+ASTPtr parseCreateStatement(const String & statement)
+{
+    ParserCreateQuery parser;
+    const char * pos = statement.data();
+    std::string error_msg;
+    auto ast = tryParseQuery(
+        parser,
+        pos,
+        pos + statement.size(),
+        error_msg,
+        /*hilite=*/false,
+        String("in ") + __PRETTY_FUNCTION__,
+        /*allow_multi_statements=*/false,
+        0);
+    if (!ast)
+        throw Exception(error_msg, ErrorCodes::SYNTAX_ERROR);
+    return ast;
+}
+
+TableID createDBAndTable(String db_name, String table_name)
+{
+    auto ctx = TiFlashTestEnv::getContext();
+    {
+        // Create database
+        const String statement = "CREATE DATABASE " + db_name + " ENGINE=TiFlash";
+        ASTPtr ast = parseCreateStatement(statement);
+        EXPECT_NE(ast, nullptr);
+        InterpreterCreateQuery interpreter(ast, *ctx);
+        interpreter.setInternal(true);
+        interpreter.setForceRestoreData(false);
+        interpreter.execute();
+    }
+
+    auto db = ctx->tryGetDatabase(db_name);
+    EXPECT_NE(db, nullptr);
+    EXPECT_EQ(db->getEngineName(), "TiFlash");
+    EXPECT_TRUE(db->empty(*ctx));
+
+    {
+        /// Create table
+        ParserCreateQuery parser;
+        const String stmt = "CREATE TABLE `" + db_name + "`.`" + table_name
+            + "`("
+              "c_custkey Int32,"
+              "c_acctbal Decimal(15, 2),"
+              "c_comment String"
+              ") ENGINE = DeltaMerge(c_custkey)";
+        ASTPtr ast = parseQuery(parser, stmt, 0);
+
+        InterpreterCreateQuery interpreter(ast, *ctx);
+        interpreter.setInternal(true);
+        interpreter.setForceRestoreData(false);
+        interpreter.execute();
+    }
+
+    EXPECT_FALSE(db->empty(*ctx));
+    EXPECT_TRUE(db->isTableExist(*ctx, table_name));
+
+    TableID table_id;
+    {
+        // Get storage from database
+        auto storage = db->tryGetTable(*ctx, table_name);
+        EXPECT_NE(storage, nullptr);
+
+        StorageDeltaMergePtr storage_ptr = std::static_pointer_cast<StorageDeltaMerge>(storage);
+        table_id = storage_ptr->getTableInfo().id;
+
+        EXPECT_EQ(storage->getName(), MutSup::delta_tree_storage_name);
+        EXPECT_EQ(storage->getTableName(), table_name);
+
+        auto managed_storage = std::dynamic_pointer_cast<IManageableStorage>(storage);
+        EXPECT_EQ(managed_storage->getDatabaseName(), db_name);
+    }
+    return table_id;
+}
+
+void dropDataBase(String db_name)
+{
+    auto ctx = TiFlashTestEnv::getContext();
+    auto drop_query = std::make_shared<ASTDropQuery>();
+    drop_query->database = db_name;
+    drop_query->if_exists = false;
+    ASTPtr ast_drop_query = drop_query;
+    InterpreterDropQuery drop_interpreter(ast_drop_query, *ctx);
+    drop_interpreter.execute();
+
+    auto db = ctx->tryGetDatabase(db_name);
+    ASSERT_EQ(db, nullptr);
+}
+
+void createRegions(size_t region_num, TableID table_id)
+{
+    auto & tmt = TiFlashTestEnv::getContext()->getTMTContext();
+    for (size_t i = 0; i < region_num; i++)
+    {
+        auto region = RegionBench::makeRegionForTable(i, table_id, i, i + region_num + 10);
+        tmt.getRegionTable().shrinkRegionRange(*region);
+    }
+}
+
+void makeRegionsLag(size_t lag_num)
+{
+    auto & tmt = TiFlashTestEnv::getContext()->getTMTContext();
+    for (size_t i = 0; i < lag_num; i++)
+    {
+        tmt.getRegionTable().safeTsMgr().updateSafeTS(
+            i,
+            (SafeTsManager::SafeTsDiffThreshold + 1) << TsoPhysicalShiftBits,
+            0);
+    }
+}
+
+TEST_F(StatusServerTest, TestSyncStatusLagRegion)
+try
+{
+    TableID table_id = createDBAndTable("db_1", "t_1");
+    // create 20 region for the table_id, but 10 regions are lagged
+    createRegions(20, table_id);
+    makeRegionsLag(10);
+    EngineStoreServerWrap store_server_wrap{};
+    store_server_wrap.tmt = &TiFlashTestEnv::getContext()->getTMTContext();
+    auto helper = GetEngineStoreServerHelper(&store_server_wrap);
+    String path = fmt::format("/tiflash/sync-status/{}", table_id);
+    auto res = helper.fn_handle_http_request(
+        &store_server_wrap,
+        BaseBuffView{path.data(), path.length()},
+        BaseBuffView{path.data(), path.length()},
+        BaseBuffView{"", 0});
+    EXPECT_EQ(res.status, HttpRequestStatus::Ok);
+    {
+        // normal region count is 10.
+        StringRef sr(res.res.view.data, res.res.view.len);
+        EXPECT_TRUE(startsWith(sr, "10\n"));
+        EXPECT_EQ(res.res.view.len, 33);
+        sr = removePrefix(sr, 3);
+        // parse the region_ids
+        std::stringstream ss(sr.toString());
+        std::set<RegionID> region_ids;
+        RegionID region_id;
+        while (ss >> region_id)
+        {
+            ASSERT_GE(region_id, 10);
+            ASSERT_LT(region_id, 20);
+            region_ids.insert(region_id);
+        }
+        ASSERT_EQ(region_ids.size(), 10) << fmt::format("{} ", region_ids);
+    }
+    releaseResp(helper, std::move(res));
+    dropDataBase("db_1");
+}
+CATCH
+
+TEST_F(StatusServerTest, TestSyncStatusNormalRegion)
+try
+{
+    TableID table_id = createDBAndTable("db_1", "t_1");
+    createRegions(20, table_id);
+    EngineStoreServerWrap store_server_wrap{};
+    store_server_wrap.tmt = &TiFlashTestEnv::getContext()->getTMTContext();
+    auto helper = GetEngineStoreServerHelper(&store_server_wrap);
+    String path = fmt::format("/tiflash/sync-status/{}", table_id);
+    auto res = helper.fn_handle_http_request(
+        &store_server_wrap,
+        BaseBuffView{path.data(), path.length()},
+        BaseBuffView{path.data(), path.length()},
+        BaseBuffView{"", 0});
+    EXPECT_EQ(res.status, HttpRequestStatus::Ok);
+    {
+        // normal region count is 20.
+        StringRef sr(res.res.view.data, res.res.view.len);
+        EXPECT_TRUE(startsWith(sr, "20\n"));
+        EXPECT_EQ(res.res.view.len, 53);
+        sr = removePrefix(sr, 3);
+        // parse the region_ids
+        std::stringstream ss(sr.toString());
+        std::set<RegionID> region_ids;
+        RegionID region_id;
+        while (ss >> region_id)
+        {
+            ASSERT_GE(region_id, 0);
+            ASSERT_LT(region_id, 20);
+            region_ids.insert(region_id);
+        }
+        ASSERT_EQ(region_ids.size(), 20) << fmt::format("{} ", region_ids);
+    }
+    releaseResp(helper, std::move(res));
+    dropDataBase("db_1");
 }
 CATCH
